Create Chunk(Range) equality tests
